Fix: Bump dependencies and various improvements
- Bump `@adbl/cells` and `@oxc-parser` dependencies - Fix: Use
contentRef to track the current sheet height for delete and edit actions
- Fix: Ensure `contentLoaded` state is reset when bottom sheet is closed
- Chore: Make sure there's no additional transition duration for
ExpandingView on exit - Fix: Don't remove the parent key of a subkey is
being removed. - Fix: Prevent the UI from becoming unmounted when
onboarding is finished - Fix: Prevent race condition from deleting a
transaction that has already been deleted - Fix: Update various
dependencies and runtime dependencies for `@oxc-parser`. - Fix: Ensure
Home Screen transitions correctly with the new `TransitionView` setup -
Fix: Use supportsCSSTypedOM to enable `attributeStyleMap` updates when
it's available.

+++ utilities/composables/use-route-query-control.ts
@@ -11,16 +11,18 @@ export function useRouteQueryControl(_key: string | { _root: string }, value = '
    const add = () => {
       query.set(key, value)
    }
-   const remove = () => {
+   const remove = (options: { subKeys?: boolean } = { subKeys: true }) => {
       const subKeys = [...current.get().query.keys()].filter((_key) => {
          return _key.startsWith(`${key}.`)
       })
       query.delete(key)
-      defer(() => {
-         // This tries to prevent any subtle timing/rendering issues
-         // that can come from trying to remove all the keys at once.
-         query.delete(...subKeys)
-      })
+      if (options.subKeys) {
+         defer(() => {
+            // This tries to prevent any subtle timing/rendering issues
+            // that can come from trying to remove all the keys at once.
+            query.delete(...subKeys)
+         })
+      }
    }
    const hasKey = Cell.derived(() => {
       return current.get().query.get(key) === value
